requirements.txt
alembic==1.4.1
anyio==3.5.0
apache-airflow==2.0.1
apache-airflow-providers-ftp==1.0.1
apache-airflow-providers-google==1.0.0
apache-airflow-providers-http==1.1.0
apache-airflow-providers-imap==1.0.1
apache-airflow-providers-postgres==1.0.1
apache-airflow-providers-sqlite==1.0.1
apache-airflow-providers-ssh==1.3.0
apispec==3.3.2
appnope==0.1.3
argcomplete==1.12.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.5.6
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bash==0.6
bcrypt==3.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
cached-property==1.5.2
cachetools==4.2.1
cattrs==1.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==8.1.2
clickclick==20.10.2
cloudpickle==1.6.0
colorama==0.4.4
colorlog==4.7.2
commonmark==0.9.1
connexion==2.7.0
croniter==0.3.37
cryptography==3.4.5
databricks-cli==0.14.3
dataclasses-json==0.5.2
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.6.0
dill==0.3.3
dnspython==1.16.0
docker==5.0.0
docutils==0.16
email-validator==1.1.2
entrypoints==0.3
fastjsonschema==2.15.3
Flask==2.1.0
Flask-AppBuilder==3.1.1
Flask-Babel==1.0.0
Flask-Caching==1.9.0
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gitdb==4.0.7
GitPython==3.1.14
google-ads==7.0.0
google-api-core==1.26.3
google-api-python-client==1.12.8
google-auth==1.29.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-automl==1.0.1
google-cloud-bigquery==2.13.1
google-cloud-bigquery-datatransfer==1.1.1
google-cloud-bigquery-storage==2.4.0
google-cloud-bigtable==1.7.0
google-cloud-container==1.0.1
google-cloud-core==1.6.0
google-cloud-datacatalog==0.7.0
google-cloud-dataproc==1.1.1
google-cloud-dlp==1.0.0
google-cloud-kms==1.4.0
google-cloud-language==1.3.0
google-cloud-logging==1.15.1
google-cloud-memcache==0.3.0
google-cloud-monitoring==1.1.0
google-cloud-os-login==1.0.0
google-cloud-pubsub==1.7.0
google-cloud-redis==1.0.0
google-cloud-secret-manager==1.0.0
google-cloud-spanner==1.19.1
google-cloud-speech==1.3.2
google-cloud-storage==1.37.1
google-cloud-tasks==1.5.0
google-cloud-texttospeech==1.0.1
google-cloud-translate==1.7.0
google-cloud-videointelligence==1.16.1
google-cloud-vision==1.0.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
graphviz==0.16
greenlet==1.0.0
grpc-google-iam-v1==0.12.3
grpcio==1.37.0
grpcio-gcp==0.2.2
gunicorn==19.10.0
httplib2==0.19.1
idna==2.10
importlib-metadata==4.11.3
importlib-resources==1.5.0
inflection==0.5.1
iniconfig==1.1.1
ipykernel==6.13.0
ipython==7.32.0
ipython-genutils==0.2.0
iso8601==0.1.14
isodate==0.6.0
isort==5.8.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
joblib==1.0.1
json-merge-patch==0.2
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.3.0
jupyter-core==4.10.0
jupyter-server==1.16.0
jupyterlab==3.3.4
jupyterlab-pygments==0.2.2
jupyterlab-server==2.13.0
lazy-object-proxy==1.4.3
libcst==0.3.18
lockfile==0.12.2
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==2.1.1
marshmallow==3.10.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.1.0
marshmallow-sqlalchemy==0.23.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mlflow==1.15.0
mypy-extensions==0.4.3
natsort==7.1.1
nbclassic==0.3.7
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
nltk==3.6.2
notebook==6.4.11
notebook-shim==0.1.0
numpy==1.20.1
oauthlib==3.1.0
openapi-spec-validator==0.2.9
packaging==20.9
pandas==1.2.2
pandas-gbq==0.15.0
pandocfilters==1.5.0
paramiko==2.7.2
parso==0.8.3
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prison==0.1.3
prometheus-client==0.10.1
prometheus-flask-exporter==0.18.1
prompt-toolkit==3.0.29
proto-plus==1.18.1
protobuf==3.15.8
psutil==5.8.0
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.10.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydata-google-auth==1.2.0
Pygments==2.8.0
PyJWT==1.7.1
pylint==2.8.2
pymystem3==0.2.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pysftp==0.2.9
pytest==6.2.3
python-daemon==2.2.4
python-dateutil==2.8.2
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python-youtube==0.8.0
python3-openid==3.2.0
pytz==2020.5
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==22.3.0
querystring-parser==1.2.4
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rich==9.2.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.2
Send2Trash==1.8.0
setproctitle==1.2.2
six==1.15.0
smmap==4.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.3.23
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.36.8
sqlparse==0.4.1
sshtunnel==0.1.5
stringcase==1.2.0
swagger-ui-bundle==0.0.8
tabulate==0.8.7
tenacity==6.2.0
termcolor==1.1.0
terminado==0.13.3
text-unidecode==1.3
threadpoolctl==2.1.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.1
tqdm==4.60.0
traitlets==5.1.1
typed-ast==1.4.3
typing-extensions==3.7.4.3
typing-inspect==0.6.0
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.0.0
wrapt==1.12.1
WTForms==2.3.3
zipp==3.4.0
